mcp_prueba/pyproject.toml
[project]
name = "prueba2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-mcp>=0.2.0",
    "ipykernel>=6.29.5",
    "linkup-sdk>=0.2.4",
    "llama-index>=0.12.30",
    "llama-index-llms-ollama>=0.5.4",
    "llama-index-tools-mcp>=0.1.1",
    "mcp[cli]>=1.6.0",
    "psycopg",
    "python-dotenv",
]